refactor(scripts): isolate script-as-CLI from script-as-module (#781)
Move the importable logic of three 'check_' scripts (check_cli_table_sync,
check_version_single_source, check_python_version_consistency) into a new
hephaestus.scripts_lib sub-package, reducing scripts/ to thin CLI entry-point
shims. Remove scripts/ from pytest.ini_options pythonpath to prevent
scripts/__pycache__ from materializing during test runs.

- Create hephaestus/scripts_lib/ with three implementation modules
- Replace scripts/check_*.py bodies with one-line shims delegating to library
- Move three test files from tests/unit/scripts/ to tests/unit/scripts_lib/
- Drop sys.path.insert hacks and import directly from hephaestus.scripts_lib
- Update monkeypatch targets to point to moved module homes
- Remove "scripts" from [tool.pytest.ini_options].pythonpath in pyproject.toml
- Clean up old test files and pycache

Verified: all 3193 unit tests pass, pre-commit hooks work, no scripts/__pycache__
materializes during tests, and no code imports from old flat script paths.
