+++ hephaestus/scripts_lib/__init__.py
@@ -1,0 +1,5 @@
+"""Importable library code for repo-meta consistency check scripts.
+
+Each module here is paired with a thin CLI shim in ``scripts/`` of the same
+name. Keep this package import-light: no eager re-exports.
+"""
